Blur capabilities separete from view controller
